+++ server/src/test/java/com/seebie/server/HistogramCalculatorTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.*;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 import static com.seebie.server.test.data.TestData.create30DayRange;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -27,13 +28,13 @@ public class HistogramCalculatorTest {
     @Test
     public void testPartialData() {
 
-        var dataSet1 = new FilterResult(create30DayRange(0), List.of(60, 120, 120));
+        var dataSet1 = new FilterResult(create30DayRange(0), List.of(60L, 120L, 120L));
         var dataSet2 = new FilterResult(create30DayRange(0), List.of());
 
         var result = calculator.buildNormalizedHistogram(60, List.of(dataSet1, dataSet2));
 
         assertEquals(2, result.bins().size());
-        assertEquals(List.of(60, 120), result.bins());
+        assertEquals(List.of(60L, 120L), result.bins());
         assertEquals(List.of(33L, 67L), result.dataSets().get(0));
         assertEquals(List.of(0L,0L), result.dataSets().get(1));
     }
@@ -45,8 +46,8 @@ public class HistogramCalculatorTest {
 
         final int binSize = req.binSize();
 
-        List<Integer> durationMinutes1 = IntStream.range(4 * 60, 8 * 60).boxed().toList();
-        List<Integer> durationMinutes2 = IntStream.range(5 * 60, 7 * 60).boxed().toList();
+        List<Long> durationMinutes1 = LongStream.range(4 * 60, 8 * 60).boxed().toList();
+        List<Long> durationMinutes2 = LongStream.range(5 * 60, 7 * 60).boxed().toList();
 
         var dataSet1 = new FilterResult(create30DayRange(0), durationMinutes1);
         var dataSet2 = new FilterResult(create30DayRange(0), durationMinutes2);
